+++ main.py
@@ -1,31 +1,6 @@
-#this is our main 
+from classes import Character, Location, Item
 
-class Character:
-    def __init__(self, name, alert_level):
-        self.name = name
-        self.alert_level = alert_level
-        self.items = []
-
-    def change_alert(self, location, item_idx):
-        self.alert_level += location.item[item_idx]
-    
-    def print_alert_status(self):
-        print(self.alert_level)
-
-    def add_items(self, item):
-        self.items.append(item)
-
-class Location:
-    def __init__(self, name, description):
-        self.name = name
-        self.description = description
-        
-
-class Item:
-    def __init__(self, name, associated_room, alert_effect=0):
-        self.name = name        
-        self.associated_room = associated_room
-        self.alert_effect = alert_effect
+# this is our main program
 
 
 # Character Instantiation
@@ -33,7 +8,7 @@ crystal = Character('Crystal', 40)
 jojo = Character('JoJo', 10)
 kurtis = Character('Kurtis', 30)
 joshua = Character('Joshua', 5)
-annalise  = Character('Annalise Keating', 0)
+annalise = Character('Annalise Keating', 0)
 
 # Location Instantiation
 josh_room = Location("Quiet Room", "pass")
@@ -53,7 +28,8 @@ items_josh_room = [monitor.name, jacket.name, bag_of_chips.name]
 
 # Liz's Room
 tshirt = Item("tshirt", liz_office, -10)
-book = Item("book", liz_office)# input code in book and if you can solve unlocks ryan phone number and he distracts security guard for you 
+# input code in book and if you can solve unlocks ryan phone number and he distracts security guard for you
+book = Item("book", liz_office)
 keyboard = Item("keyboard", liz_office, 15)
 items_liz_room = [tshirt.name, book.name, keyboard.name]
 
@@ -87,11 +63,12 @@ items_security_desk = [security_desk.name]
 
 # Menus
 
-main_menu = ["Search for anything useful", "Look at my items", "Move to a new location", "Call the Police"]
-
+main_menu = ["Search for anything useful", "Look at my items",
+             "Move to a new location", "Call the Police"]
 
 
 main_players = [crystal.name, jojo.name, kurtis.name, joshua.name]
+
 
 def which_list(room):
     if room == josh_room:
@@ -103,8 +80,8 @@ def which_list(room):
 def play_game():
 
     print("Welcome to How To Get Away With Murder")
-# We need some info here about the game....so you know what character you want to choose, must use items in Josh's room!
-    print() #Add in player stats and special characteristics
+    # We need some info here about the game....so you know what character you want to choose, must use items in Josh's room!
+    print()  # Add in player stats and special characteristics
     for idx, player in enumerate(main_players):
         print(f"{idx+1}. {player}")
     choosing_player = input("Please pick a player: ")
@@ -119,8 +96,9 @@ def play_game():
     if choosing_player == "4":
         active_player = joshua
     if choosing_player == "1000":
-        active_player = annalise    
-    print(f'You have selected {active_player.name} with an alert level of {active_player.alert_level}!')
+        active_player = annalise
+    print(
+        f'You have selected {active_player.name} with an alert level of {active_player.alert_level}!')
     # Here we need to set the scene!
     curr_location = josh_room
     curr_items_list = which_list(curr_location)
@@ -138,18 +116,14 @@ def play_game():
             item_chosen = int(input("Which item would you like to pick up? "))
             active_player.add_items(curr_items_list[item_chosen-1])
             curr_items_list.pop(item_chosen-1)
-            continue_choosing = input("Would you like to choose another item? y or n\n")
+            continue_choosing = input(
+                "Would you like to choose another item? y or n\n")
             if continue_choosing.lower() == "n":
                 item_loop = False
-
-       
 
     # if main_menu_choice == "2":
     # if main_menu_choice == "3":
     # if main_menu_choice == "4":
-      
 
 
 play_game()
-
-
